// client/libs/lib-redux/src/lib/redux-store/ui/active-project/active-project.models.ts
export interface Panel {
  id: number;
  tabs: PanelTab<any>[];
}

export type ListType = '' | 'digitals' | 'entities' | 'sources' | 'analysis' | 'queries' | 'visuals' | 'stories' | 'settings';
export interface PanelTab<D> {
  // wheter tab is active or not
  active: boolean;
  // the root component included in this tab, in dash separate minuscles: PeItDetailComponent -> 'entity-detail'
  component: LbClassConfig.DetailPageEnum;
  // icon to be displayed in tab, e.g.: gv-icon-source
  icon: LbClassConfig.IconEnum,
  // data to pass to component via input variabales
  data?: D
  // generated by reducer: id of the tab
  id?: string;
  // generated on the fly, never in store
  tabTitle$?: Observable<string>;
  loading$?: Observable<boolean>;
}